fix QR scanning
